requirements.txt
django>=5.2
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.3
pillow>=10.0
python-dotenv>=1.0
stripe>=8.0
boto3>=1.34